requirements.txt
python-binance==1.0.15
fastapi==0.75.1
pandas==1.4.1
pytest-asyncio==0.18.3
